Remove deployment client
